Move functions to conf.d
There's a compromise at play here. Keeping the functions in `functions`
means they will be conveniently automatically reloaded whenever they're
edited, but at the cost that each function must be defined in its own
file, which is painful in the NixOS config. Storing the functions in
`conf.d` allows consolidation in a few files, which makes for easier
NixOS maintenance at the cost of losing the automatic reload capability.
I opt for the latter.